Refactoring hdf5tree into package silx.gui.hdf5
- Fix warnings
