+++ silx/gui/hdf5/__init__.py
@@ -22,26 +22,16 @@
 # THE SOFTWARE.
 #
 # ###########################################################################*/
-__authors__ = ["T. Vincent"]
+"""This package provides a set of Qt widgets for displaying content relative to
+HDF5 format.
+"""
+
+__authors__ = ["V. Valls"]
 __license__ = "MIT"
 __date__ = "13/09/2016"
 
 
-from numpy.distutils.misc_util import Configuration
+from .Hdf5TreeView import Hdf5TreeView  # noqa
+from ._utils import Hdf5ContextMenuEvent  # noqa
 
-
-def configuration(parent_package='', top_path=None):
-    config = Configuration('gui', parent_package, top_path)
-    config.add_subpackage('plot')
-    config.add_subpackage('fit')
-    config.add_subpackage('hdf5')
-    config.add_subpackage('widgets')
-    config.add_subpackage('test')
-
-    return config
-
-
-if __name__ == "__main__":
-    from numpy.distutils.core import setup
-
-    setup(configuration=configuration)
+__all__ = ['Hdf5TreeView', 'Hdf5ContextMenuEvent']
